Add sidebar navigation and hamburger icon functionality
A sidebar navigation has been added to the dashboard.html file which includes links to Dashboard, Tasks, Notes and settings. Also, functionality has been added to the hamburger icon so that it can be used to toggle the visibility of this new sidebar. This drastically improves the application's navigation and usability.