Add data export functionality
A new route '/export' has been added to enable users to export their notes and tasks in a JSON format. This was accomplished by querying the user's notes and tasks, and writing the resulting data into a 'export.json' file, which is then sent as a downloadable file to the user. Required dependencies were also imported.